+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **core**: BREAKING - `diff` is now async
 - **core**: Adds `considerUsage` rule
 - **core**: Fixes missing names of default root types
+- **cli**, **ci**: Adds `@aws_lambda` directive
 
 ### v2.9.0